findstore/requirements.txt
branca==0.3.1
certifi==2019.11.28
chardet==3.0.4
cycler==0.10.0
defusedxml==0.6.0
Django==2.2.7
django-allauth==0.42.0
django-cors-headers==3.3.0
django-rest-auth==0.9.5
djangorestframework==3.10.3
folium==0.10.1
idna==2.8
Jinja2==2.10.3
joblib==0.14.0
kiwisolver==1.1.0
MarkupSafe==1.1.1
matplotlib==3.1.2
numpy==1.17.3
oauthlib==3.1.0
pandas==0.25.3
pyparsing==2.4.5
python-dateutil==2.8.1
python3-openid==3.2.0
pytz==2019.3
requests==2.22.0
requests-oauthlib==1.3.0
scikit-learn==0.21.3
scipy==1.3.1
seaborn==0.9.0
six==1.13.0
sqlparse==0.3.0
urllib3==1.25.7
